replace old fetcher with the new one
