+++ src/main/java/org/jabref/logic/importer/fetcher/AstrophysicsDataSystem.java
@@ -1,10 +1,16 @@
 package org.jabref.logic.importer.fetcher;
 
+import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URISyntaxException;
 import java.net.URL;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.jabref.logic.cleanup.MoveFieldCleanup;
 import org.jabref.logic.formatter.bibtexfields.ClearFormatter;
@@ -16,16 +22,22 @@ import org.jabref.logic.importer.EntryBasedParserFetcher;
 import org.jabref.logic.importer.FetcherException;
 import org.jabref.logic.importer.IdBasedParserFetcher;
 import org.jabref.logic.importer.ImportFormatPreferences;
+import org.jabref.logic.importer.ParseException;
 import org.jabref.logic.importer.Parser;
 import org.jabref.logic.importer.SearchBasedParserFetcher;
 import org.jabref.logic.importer.fileformat.BibtexParser;
+import org.jabref.logic.net.URLDownload;
 import org.jabref.model.cleanup.FieldFormatterCleanup;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.StandardField;
 import org.jabref.model.entry.field.UnknownField;
+import org.jabref.model.strings.StringUtil;
 import org.jabref.model.util.DummyFileUpdateMonitor;
 
 import org.apache.http.client.utils.URIBuilder;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
 
 /**
  * Fetches data from the SAO/NASA Astrophysics Data System (http://www.adsabs.harvard.edu/)
@@ -37,15 +49,24 @@ import org.apache.http.client.utils.URIBuilder;
  */
 public class AstrophysicsDataSystem implements IdBasedParserFetcher, SearchBasedParserFetcher, EntryBasedParserFetcher {
 
-    private static String API_QUERY_URL = "http://adsabs.harvard.edu/cgi-bin/nph-basic_connect";
-    private static String API_ENTRY_URL = "http://adsabs.harvard.edu/cgi-bin/nph-abs_connect";
-    private static String API_DOI_URL = "http://adsabs.harvard.edu/doi/";
+    private static final String API_SEARCH_URL = "https://api.adsabs.harvard.edu/v1/search/query";
+    private static final String API_EXPORT_URL = "https://api.adsabs.harvard.edu/v1/export/bibtexabs";
 
-    private final String patternRemoveDOI = "^(doi:|DOI:)";
+    private static final String API_KEY = ""; //TODO Add API Token
     private final ImportFormatPreferences preferences;
 
     public AstrophysicsDataSystem(ImportFormatPreferences preferences) {
         this.preferences = Objects.requireNonNull(preferences);
+    }
+
+    private String buildPostData(Collection<String> bibcodes) {
+        JSONObject obj = new JSONObject();
+        obj.put("bibcode", bibcodes);
+        return obj.toString();
+    }
+
+    private URL getURLforExport() throws URISyntaxException, MalformedURLException {
+        return new URIBuilder(API_EXPORT_URL).build().toURL();
     }
 
     @Override
@@ -53,57 +74,44 @@ public class AstrophysicsDataSystem implements IdBasedParserFetcher, SearchBased
         return "SAO/NASA Astrophysics Data System";
     }
 
-    private URIBuilder getBaseUrl(String apiUrl) throws URISyntaxException {
-        URIBuilder uriBuilder = new URIBuilder(apiUrl);
-        uriBuilder.addParameter("data_type", "BIBTEXPLUS");
-        uriBuilder.addParameter("start_nr", String.valueOf(1));
-        uriBuilder.addParameter("nr_to_return", String.valueOf(200));
-        return uriBuilder;
-    }
-
     @Override
     public URL getURLForQuery(String query) throws URISyntaxException, MalformedURLException, FetcherException {
-        URIBuilder uriBuilder = getBaseUrl(API_QUERY_URL);
-        uriBuilder.addParameter("qsearch", query);
-        return uriBuilder.build().toURL();
+        URIBuilder builder = new URIBuilder(API_SEARCH_URL);
+        builder.addParameter("q", query);
+        builder.addParameter("fl", "bibcode");
+        return builder.build().toURL();
     }
 
     @Override
     public URL getURLForEntry(BibEntry entry) throws URISyntaxException, MalformedURLException, FetcherException {
-        URIBuilder uriBuilder = getBaseUrl(API_ENTRY_URL);
+        StringBuilder stringBuilder = new StringBuilder();
 
-        // Search astronomy + physics + arXiv db
-        uriBuilder.addParameter("db_key", "AST");
-        uriBuilder.addParameter("db_key", "PHY");
-        uriBuilder.addParameter("db_key", "PRE");
+        Optional<String> title = entry.getFieldOrAlias(StandardField.TITLE).map(t -> "title:\"" + t + "\"");
+        Optional<String> author = entry.getFieldOrAlias(StandardField.AUTHOR).map(a -> "author:\"" + a + "\"");
 
-        // Add title search
-        entry.getFieldOrAlias(StandardField.TITLE).ifPresent(title -> {
-            uriBuilder.addParameter("ttl_logic", "OR");
-            uriBuilder.addParameter("title", title);
-            uriBuilder.addParameter("ttl_syn", "YES"); // Synonym replacement
-            uriBuilder.addParameter("ttl_wt", "0.3"); // Weight
-            uriBuilder.addParameter("ttl_wgt", "YES"); // Consider Weight
-        });
+        if (title.isPresent()) {
+            stringBuilder.append(title.get())
+                         .append(author.map(s -> " AND " + s)
+                                       .orElse(""));
+        } else {
+            stringBuilder.append(author.orElse(""));
+        }
+        String query = stringBuilder.toString().trim();
 
-        // Add author search
-        entry.getFieldOrAlias(StandardField.AUTHOR).ifPresent(author -> {
-            uriBuilder.addParameter("aut_logic", "OR");
-            uriBuilder.addParameter("author", author);
-            uriBuilder.addParameter("aut_syn", "YES"); // Synonym replacement
-            uriBuilder.addParameter("aut_wt", "1.0"); // Weight
-            uriBuilder.addParameter("aut_wgt", "YES"); // Consider weight
-        });
-
-        return uriBuilder.build().toURL();
+        URIBuilder builder = new URIBuilder(API_SEARCH_URL);
+        builder.addParameter("q", query);
+        builder.addParameter("fl", "bibcode");
+        builder.addParameter("rows", "20");
+        return builder.build().toURL();
     }
 
     @Override
-    public URL getURLForID(String identifier) throws URISyntaxException, MalformedURLException, FetcherException {
-        String key = identifier.replaceAll(patternRemoveDOI, "");
-        URIBuilder uriBuilder = new URIBuilder(API_DOI_URL + key);
-        uriBuilder.addParameter("data_type", "BIBTEXPLUS");
-        return uriBuilder.build().toURL();
+    public URL getURLForID(String identifier) throws FetcherException, URISyntaxException, MalformedURLException {
+        String query = "doi:\"" + identifier + "\" OR " + "bibcode:\"" + identifier + "\"";
+        URIBuilder builder = new URIBuilder(API_SEARCH_URL);
+        builder.addParameter("q", query);
+        builder.addParameter("fl", "bibcode");
+        return builder.build().toURL();
     }
 
     @Override
@@ -129,5 +137,124 @@ public class AstrophysicsDataSystem implements IdBasedParserFetcher, SearchBased
         new MoveFieldCleanup(new UnknownField("adsurl"), StandardField.URL).cleanup(entry);
         // The fetcher adds some garbage (number of found entries etc before)
         entry.setCommentsBeforeEntry("");
+    }
+
+    @Override
+    public List<BibEntry> performSearch(BibEntry entry) throws FetcherException {
+
+        if (entry.getFieldOrAlias(StandardField.TITLE).isEmpty() && entry.getFieldOrAlias(StandardField.AUTHOR).isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        try {
+            List<String> bibcodes = fetchBibcodes(getURLForEntry(entry));
+            return performSearchByIds(bibcodes);
+        } catch (URISyntaxException e) {
+            throw new FetcherException("Search URI is malformed", e);
+        } catch (IOException e) {
+            throw new FetcherException("A network error occurred", e);
+        }
+    }
+
+    @Override
+    public List<BibEntry> performSearch(String query) throws FetcherException {
+
+        if (StringUtil.isBlank(query)) {
+            return Collections.emptyList();
+        }
+
+        try {
+            List<String> bibcodes = fetchBibcodes(getURLForQuery(query));
+            return performSearchByIds(bibcodes);
+        } catch (URISyntaxException e) {
+            throw new FetcherException("Search URI is malformed", e);
+        } catch (IOException e) {
+            throw new FetcherException("A network error occurred", e);
+        }
+    }
+
+    private List<String> fetchBibcodes(URL url) throws FetcherException {
+
+        try {
+            URLDownload download = new URLDownload(url);
+            download.addHeader("Authorization", "Bearer " + API_KEY);
+            String content = download.asString();
+            JSONObject obj = new JSONObject(content);
+
+            try {
+                JSONArray codes = obj.getJSONObject("response").getJSONArray("docs");
+                List<String> bibcodes = new ArrayList<>();
+                for (int i = 0; i < codes.length(); i++) {
+                    bibcodes.add(codes.getJSONObject(i).getString("bibcode"));
+                }
+                return bibcodes;
+            } catch (JSONException e) {
+                return Collections.emptyList();
+            }
+        } catch (IOException e) {
+            throw new FetcherException("A network error occurred", e);
+        }
+    }
+
+    @Override
+    public Optional<BibEntry> performSearchById(String identifier) throws FetcherException {
+        if (StringUtil.isBlank(identifier)) {
+            return Optional.empty();
+        }
+
+        try {
+            List<String> bibcodes = fetchBibcodes(getURLForID(identifier));
+            List<BibEntry> fetchedEntries = performSearchByIds(bibcodes);
+
+            if (fetchedEntries.isEmpty()) {
+                return Optional.empty();
+            }
+            if (fetchedEntries.size() > 1) {
+                LOGGER.info("Fetcher " + getName() + "found more than one result for identifier " + identifier
+                        + ". We will use the first entry.");
+            }
+            BibEntry entry = fetchedEntries.get(0);
+            return Optional.of(entry);
+        } catch (URISyntaxException e) {
+            throw new FetcherException("Search URI is malformed", e);
+        } catch (IOException e) {
+            throw new FetcherException("A network error occurred", e);
+        }
+    }
+
+    private List<BibEntry> performSearchByIds(Collection<String> identifiers) throws FetcherException {
+
+        List<String> ids = identifiers.stream().filter(identifier -> !StringUtil.isBlank(identifier)).collect(Collectors.toList());
+        if (ids.isEmpty()) {
+            return Collections.emptyList();
+        }
+        try {
+            String postData = buildPostData(ids);
+            URLDownload download = new URLDownload(getURLforExport());
+            download.addHeader("Authorization", "Bearer " + API_KEY);
+            download.addHeader("ContentType", "application/json");
+            download.setPostData(postData);
+            String content = download.asString();
+            JSONObject obj = new JSONObject(content);
+
+            try {
+                List<BibEntry> fetchedEntries = getParser().parseEntries(obj.optString("export"));
+                if (fetchedEntries.isEmpty()) {
+                    return Collections.emptyList();
+                }
+                // Post-cleanup
+                fetchedEntries.forEach(this::doPostCleanup);
+
+                return fetchedEntries;
+            } catch (JSONException e) {
+                return Collections.emptyList();
+            }
+        } catch (URISyntaxException e) {
+            throw new FetcherException("Search URI is malformed", e);
+        } catch (IOException e) {
+            throw new FetcherException("A network error occurred", e);
+        } catch (ParseException e) {
+            throw new FetcherException("An internal parser error occurred", e);
+        }
     }
 }
